getPlayerSelection began and playRound functional
Previous iteration had only 'let playerSelection = "Rock"';, this was
changed to include the new function 'let playerSelection =
getPlayerSelection();'. This function offers a pop out prompt and
receives a user input. This is still case sensitive, however. Tomorrow
I need to test the function with .toUpperCase or .toLowerCase.
This will make the function case-insensitive, but will require me to
redo the "function getComputerChoice()" result to "result.toUpperCase"
and choice to "choice.toUpperCase".

The playRound function works! It still occasionally returns "undefined".
This is most likely due to the math in "getComputerChoice()". Will
work on it tomorrow as well.