javascript2 fixes and comments
TOP discord user @HappySlappyFace tweaked my work and gave me plenty of
notes. Javascript2 is his fixed version of my file with citations.

I added his notes, and tested them against the game. It worked much
better! I expanded back out to 5 games, and found a new bug: the
game ticks up the first time for a win, but it never ticks up for
computerScore, and it stops after one increment up for playerScore.

Probably an issue with the ++ model, and the fact that I am not looping.