game() finished for now
There were two more bugs I was working on:

1. game does not tick up computerScore or playerScore.
2. computer gives the same option over and over.

Number 2 was solved by adding computerSelection = getComputerChoice();
underneath the playerSelection = getPlayerSelection();. It wasn't giving
a new computer choice each turn, but reusing the original. This forces
the computer to choose newly each turn.

Number 1 I think will be solved with a loop, which I learn next lesson.
So I think it's okay to leave it for right now, since we will come back
and fix it later.